+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 - Added a missing package to the docker images to fix errors regarding authentication
 - Upgraded docker images to Ubuntu 26.04 resolute
 
+### Additions
+- Added a new command `azuracast get-station-logs`
+  - This command is similar to `azuracast get-system-logs` but only shows the logs of a specific station
+
 ### Improvements
 - Re-enabled HTTP/3 support after implementing several packages in the Dockerfiles
 - The bot now also sends a welcome notification to the first channel it can see when it's added to a server